@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "start": "node server.js",
-    "start:dev": "nodemon server.js",
+    "start:dev": "nodemon --ignore 'public/*' server.js",
     "start:prod": "NODE_ENV=production nodemon server.js",
     "debug": "ndb server.js",
     "build": "webpack --config webpack.config.js",
@@ -32,17 +32,25 @@
     "connect-mongo": "^3.2.0",
     "cookie-parser": "^1.4.4",
     "cors": "^2.8.5",
+    "datatables.net": "^1.10.24",
+    "datatables.net-bs4": "^1.10.24",
+    "datatables.net-colreorder": "^1.5.3",
+    "datatables.net-colreorder-bs4": "^1.5.3",
     "dotenv": "^7.0.0",
+    "expose-loader": "^2.0.0",
     "express": "^4.16.4",
     "express-mongo-sanitize": "^1.3.2",
     "express-rate-limit": "^3.5.0",
     "express-session": "^1.17.1",
+    "file-loader": "^6.2.0",
     "helmet": "^4.1.0",
     "hpp": "^0.2.2",
     "html-to-text": "^5.1.1",
+    "html-webpack-plugin": "^5.3.1",
     "jquery": "^3.5.1",
+    "jquery-datatables-checkboxes": "^1.2.12",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.0",
+    "mongoose": "^5.12.3",
     "mongoose-beautiful-unique-validation": "^7.1.1",
     "morgan": "^1.9.1",
     "nodemailer": "^6.4.14",
@@ -55,6 +63,7 @@
     "passport-oauth2-client-password": "^0.1.2",
     "perfect-scrollbar": "1.5.0",
     "pug": "^2.0.3",
+    "selectize": "^0.12.6",
     "uuid": "^8.3.0",
     "validator": "^10.11.0",
     "xss-clean": "^0.1.1"
@@ -82,7 +91,7 @@
     "sass-loader": "^10.0.3",
     "style-loader": "^2.0.0",
     "terser-webpack-plugin": "^5.0.3",
-    "webpack": "^5.1.3",
+    "webpack": "^5.33.2",
     "webpack-cli": "^4.1.0"
   },
   "engines": {
