@@ -87,19 +87,12 @@ exports.deleteMe = catchAsync(async (req, res, next) => {
   });
 });
 
-exports.createUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not defined! Please use /signup instead'
-  });
-};
-
-// exports.getUser = factory.getOne(User);
-// exports.getAllUsers = factory.getAll(User);
-
+exports.getUser = factory.getOne(User);
+exports.getAllUsers = factory.getAll(User);
 // Do NOT update passwords with this!
-// exports.updateUser = factory.updateOne(User);
-// exports.deleteUser = factory.deleteOne(User);
+exports.createUser = factory.createOne(User);
+exports.updateUser = factory.updateOne(User);
+exports.deleteUser = factory.deleteOne(User);
 
 // manual approach
 exports.find = async id => {
