@@ -22,6 +22,12 @@ const userSchema = new mongoose.Schema({
     unique: 'The username ({VALUE}) already in use. Please use a different username.',
     lowercase: true
   },
+  institute: {
+    type: String
+  },
+  lab: {
+    type: String
+  },
   photo: {
     type: String,
     default: 'default.jpg'
@@ -41,20 +47,57 @@ const userSchema = new mongoose.Schema({
   },
   password: {
     type: String,
-    required: [true, 'Please provide a password'],
-    minlength: 6,
+    validate: {
+      validator: function(el) {
+        let update;
+        if (this.getUpdate) update = this.getUpdate();
+        let loginType;
+        if (this.loginType) {
+          // for createNewField
+          loginType = this.loginType;
+        } else if (update && update['$set'] && update['$set'].loginType) {
+          loginType = update['$set'].loginType;
+        } else if (this.r && this.r.loginType) {
+          // for findByIdAndUpdate
+          loginType = this.r.loginType;
+        }
+        if (loginType == 'password') {
+          return el.length > 5;
+        }
+        return true;
+      },
+      message: 'Please provide a password longer than 5 characters'
+    },
     select: false
   },
   passwordConfirm: {
     type: String,
-    required: [true, 'Please confirm your password'],
     validate: {
-      // This only works on CREATE and SAVE!!!
       validator: function(el) {
-        return el === this.password;
+        let update;
+        if (this.getUpdate) update = this.getUpdate();
+        let loginType;
+        let password;
+        if (this.loginType) {
+          // for createNewField
+          loginType = this.loginType;
+          password = this.password;
+        } else if (update && update['$set'] && update['$set'].loginType) {
+          loginType = update['$set'].loginType;
+          password = update['$set'].password;
+        } else if (this.r && this.r.loginType) {
+          // for findByIdAndUpdate
+          loginType = this.r.loginType;
+          password = this.r.password;
+        }
+        if (loginType == 'password') {
+          return el === password;
+        }
+        return true;
       },
       message: 'Passwords are not the same!'
-    }
+    },
+    select: false
   },
   emailConfirm: {
     type: String,
@@ -106,6 +149,11 @@ userSchema.pre('save', function(next) {
 userSchema.pre(/^find/, function(next) {
   // this points to the current query
   this.find({ active: { $ne: false } });
+  next();
+});
+
+userSchema.pre(/^findOneAnd/, async function(next) {
+  this.r = await this.findOne();
   next();
 });
 
